Moved item logic to a separate file
